Your provided commit message does not match the code differences mentioned above. A suitable commit message for the code difference provided would be:
Change file retrieval method in GeneratesPDF trait

The method of retrieving a user's property contract PDF has been changed. Instead of using the Storage::path function, the Storage::url method is now being used to retrieve the file. This results in a URL instead of a directory path, which can be more useful in certain contexts.